feat: 질문 API 답변없는 질문 삭제 구현

+++ src/question/question.service.ts
@@ -95,7 +95,8 @@ export class QuestionService {
         HttpMessage.FAIL_DELETE_QUESTION,
       );
     } finally {
-      await this.answerRepository.deleteOneById(question.answer.id);
+      if (question.answer)
+        await this.answerRepository.deleteOneById(question.answer.id);
     }
 
     return;
